+++ src/App.css
@@ -1,5 +1,5 @@
 #root {
-    margin: 10px;
+    margin: 0px 10px;
     width: 100%;
 }
