+++ src/index.css
@@ -4,7 +4,7 @@
   font-weight: 400;
 
   color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
+  color: #FED27E;
   background-color: #242424;
 
   user-select: none;
