+++ src/styles/Homepage.css
@@ -13,7 +13,7 @@ body{
 .createResumeButton {
     display: inline-block;
     padding: 10px 20px;
-    background-color: #007bff;
+    background-color: rgb(107, 156, 219);
     color: #fff;
     text-decoration: none;
     border-radius: 5px;
@@ -22,7 +22,8 @@ body{
 }
 
 .createResumeButton:hover {
-    background-color: #0056b3;
+    background-color: #007bff;
+    color: white;
 }
 
 #resumeImage{
